refactor: split nativeBuildJob type and methods to a separate file
